Tests working again more
